+++ programs/ydelta/src/program/instruction_builders/place_order_for_sub_vault_instruction.rs
@@ -24,7 +24,7 @@ pub fn place_order_for_sub_vault_instruction(
     market: &Pubkey,
     fee_payer: &Pubkey,
     curator: &Pubkey,
-    sub_vault_id: u8,
+    sub_vault_id: u16,
     rate_bps: u16,
     term_seconds: u32,
     flags: u8,
